Using celery to add and updated feeds asyncronously when possible.
Based on the work by @MissFilly for Pyar Web:
PyAr/pyarweb/8e2c0e3738dafb6f4c82304ba593fae26722c055 and
PyAr/pyarweb/28681e855b2f4c9361a6daec33c2f1237c36a28d

 * Now celery and django-celery are requirements.
 * This closes #33.
 * Adding @MissFilly to AUTHORS file.
