+++ AUTHORS
@@ -4,10 +4,11 @@ The PRIMARY AUTHORS are:
 
 ADDITIONAL CONTRIBUTORS include:
 
-    * Jérémy Bethmont 
+    * Jérémy Bethmont
     * Steve Schwarz
     * Milton Mazzarri
     * @chrys
     * @akariv
     * Davide Setti
     * Joeri Poesen
+    * @MissFilly
