+++ planet/management/commands/planet_update_all_feeds.py
@@ -1,9 +1,7 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
-from datetime import datetime
 from django.core.management.base import NoArgsCommand
 
-from planet.management.commands import process_feed
+from planet.tasks import process_feed
 from planet.models import Feed
 from planet.signals import feeds_updated
 
@@ -12,12 +10,9 @@ class Command(NoArgsCommand):
     help = "Update all feeds"
 
     def handle(self, *args, **options):
-        new_posts_count = 0
-        start = datetime.now()
         for feed_url in Feed.site_objects.all().values_list("url", flat=True):
             # process feed in create-mode
-            new_posts_count += process_feed(feed_url, create=False)
-        delta = datetime.now() - start
-        print("Added {} posts in {} seconds".format(new_posts_count, delta.seconds))
-        feeds_updated.send(sender=self, instance=self)
+            self.stdout.write("Scheduling feed with URL=%s..." % feed_url)
+            process_feed.delay(feed_url, create=False)
 
+        feeds_updated.send(sender=self, instance=self)
